--- server/service/odp/createOdp.ts
@@ -1,3 +1,0 @@
-import { OdpRepository } from '@server/repository/odp'
-
-export const { createOdp } = OdpRepository
